src/requirements.txt
aiohttp==3.8.3
pydantic==1.10.4
telethon==1.24
python-dotenv==0.21.0
